Move AuthCallback into components folder
